requirements.txt
# --- Streamlit App & Core AI ---
streamlit
sentence-transformers>=2.7.0
torch
PyPDF2
docx2txt
python-docx
fpdf
requests
beautifulsoup4

# --- LangChain/Ollama Chatbot ---
langchain>=0.2.0
langchain-community>=0.2.0
langchain-text-splitters>=0.2.0
chromadb>=0.5.0
ollama>=0.2.0

# --- Notifications & Utilities ---
python-dotenv>=1.0.0
twilio>=8.0.0
python-telegram-bot>=20.0

# --- General Dependencies ---
SQLAlchemy
PyYAML